initial work on dynamic resource types
Prole now knows nothing about how to fetch particular resource types (i.e. git
vs. raw). Instead, it is configured with a mapping of (type) => (image). When
fetching a build's source of a given type, a container is spun up with the
configured image, and an executable `/tmp/resource/in` is executed with a path
to put the bits as the first argument, and the resource's version passed on
stdin.

A resource's version is a black box to prole. It comes from upstream and is
passed along to the resource. The plan is for this to ultimately come from the
same resource, as the output of `/tmp/resource/check`, which is what will be
executed to check for new versions of the resource.

Signed-off-by: dmitriy kalinin <dmitriy@pivotallabs.com>