a build can have outputs specified
A build output is evaluated after it succeeds, and specifies resulting sources
of a build. For example, an output can be to push to a git repository, which
results in a source identifying the pushed ref/branch.

The outputs are evaluated in parallel, as they should be independent from one
another.