don't pass any particular resource to /out
an output should instead be configured (via params) with the location of the
source to use. this is already done by most outputs, for example docker-image
is told either an image to build or a repository to pull down. so, there's no
current connection between an output and its 'inputs'.